Refactor code, test linear-nn on 3 datasets
