+++ results/AER/parameters.md
@@ -1,0 +1,16 @@
+### Parameter containing:
+
+```python
+tensor([[ 2.6158e-04, -2.8435e+00,  8.3571e-03, -3.5960e-01,  8.7732e-01,
+          1.6830e+00,  2.3615e-01, -1.8094e-01, -6.9595e-01, -6.5222e-03,
+         -2.9398e-02,  7.7538e-02]], dtype=torch.float64, requires_grad=True)
+```
+
+
+
+### Parameter containing:
+
+```python
+tensor([-0.1975], dtype=torch.float64, requires_grad=True)
+```
+
