+++ results/UCI/parameters.md
@@ -1,0 +1,16 @@
+### Parameter containing:
+
+```python
+tensor([[-0.0093, -0.2615, -0.0975, -0.0762, -0.0665,  0.1308, -0.1011,  0.1149,
+         -0.0889,  0.0968,  0.0966, -0.2229, -0.2180, -0.1251,  0.1090, -0.2504,
+         -0.2140,  0.0520,  0.0127,  0.0586,  0.0398, -0.1173, -0.2496]],
+       requires_grad=True)
+```
+
+
+
+### Parameter containing:
+
+```python
+tensor([-0.1054], requires_grad=True)
+```
