+++ results/clean_data/parameters.md
@@ -1,0 +1,15 @@
+### Parameter containing:
+```python
+tensor([[-2.7129e-04, -2.7406e+00,  5.3957e-03, -3.6472e-01,  1.5143e+00,
+          2.4845e-01, -1.5538e-01, -6.1257e-01,  1.1185e-01,  6.9775e-02]],
+       dtype=torch.float64, requires_grad=True)
+```
+
+
+
+### Parameter containing:
+
+```python
+tensor([-0.5944], dtype=torch.float64, requires_grad=True)
+```
+
